Prepare an ernest_sampler object for nested sampling

Create a new set of live points/particles for a new nested sampling run, or check the current state of the live points before continuing a previous run.

Usage

# S3 method for class 'ernest_sampler'
compile(object, ..., seed = NA, clear = FALSE)

Arguments

object

(ernest_sampler) An object of class ernest_sampler.

...

These dots are for future extensions and must be empty.

seed

(integer or NA, optional) Specification for the random number generator.

integer: Passed to set.seed().

NULL: Passed to set.seed(), which reinitializes the generator as if no seed has yet been set.

NA: Make no changes to the current seed if set. If compile has been called on object before, then NA will ensure that the seed remain identical between runs.

clear

(boolean) Whether to reset the sampler before compiling.

TRUE: Previous results stored in object are removed, and live points are generated and validated.

FALSE: Previous results stored in object are retained, and live points are validated.

Value

object, invisibly.

Details

The compile function prepares an ernest_sampler object for nested sampling by ensuring that its set of live points is valid and ready for use. In addition to constructing the live point set for new runs or when clear = FALSE, compile also ensures that:

The live points are each represented within the unit hypercube.

The wrapped likelihood function ernest_likelihood has a valid return value for each point (either a finite double or -Inf).

The live points don't represent a perfect plateau (i.e., all points share the same likelihood). You are warned if more than 25% of the points share the same likelihood value.

If compile fails these validation steps, the set of live points will be removed from object, preventing you from calling generate() on a malformed sampler.

Random number generation can be seeded either through the seed argument, or by calling set.seed() before running compile or generate.

Examples

prior <- create_uniform_prior(n_dim = 2, lower = -1, upper = 1)
ll_fn <- function(x) -sum(x^2)
sampler <- nested_sampling(ll_fn, prior, n_point = 100)

# Add live points to the sampler
compile(sampler)
head(sampler$live_points$unit)
#>           [,1]      [,2]
#> [1,] 0.8092600 0.8933744
#> [2,] 0.5632866 0.3311443
#> [3,] 0.4600272 0.3106995
#> [4,] 0.2293382 0.9517439
#> [5,] 0.9851504 0.3984678
#> [6,] 0.4691504 0.1493300

# Check the status of the sampler with `clear = FALSE`
compile(sampler, clear = FALSE)
head(sampler$live_points$unit)
#>           [,1]      [,2]
#> [1,] 0.8092600 0.8933744
#> [2,] 0.5632866 0.3311443
#> [3,] 0.4600272 0.3106995
#> [4,] 0.2293382 0.9517439
#> [5,] 0.9851504 0.3984678
#> [6,] 0.4691504 0.1493300

# Reset the sampler with new live points with `clear = TRUE`
compile(sampler, clear = TRUE)
head(sampler$live_points$unit)
#>           [,1]      [,2]
#> [1,] 0.7364245 0.9222809
#> [2,] 0.4591469 0.5981844
#> [3,] 0.7512136 0.4839136
#> [4,] 0.0754709 0.9646164
#> [5,] 0.1202378 0.9849806
#> [6,] 0.8002604 0.3736118