Specify a prior with Student's t-distributed marginals

A specialization of create_prior(), where the parameter space is described by independently distributed Student's t marginals that are possibly truncated.

Usage

create_t_prior(
  n_dim,
  df,
  mu = 0,
  sigma = 1,
  varnames = NULL,
  lower = -Inf,
  upper = Inf
)

Arguments

n_dim

(positive integer) The dimensionality of the prior distribution.

df

degrees of freedom (\(> 0\), maybe non-integer). df = Inf is allowed.

mu

The location parameter of the distribution. If ncp == 0 (or NULL), this is the median.

sigma

The scale parameter of the distribution.

varnames

(optional character vector) A character vector of names for the variables in the prior distribution.

lower, upper

(optional numeric vector) The values to retain from a truncated distribution.

Value

A student_t_prior, which is a subclass of ernest_prior with an efficient implementation of the unit hypercube transformation.

Details

The parameters used to specify this prior are for the location-scale t-distribution with location \(\mu\), scale \(\sigma\), and degrees of freedom \(\nu\). If the random variable \(X\) is Student-T distributed, then $$X \sim \mu + \sigma T$$ where \(T\) follows the Student-T distribution with \(\nu\) degrees of freedom.

References

For the truncation routine: Nadarajah, S., & Kotz, S. (2006). R Programs for Truncated Distributions. Journal of Statistical Software, Code Snippets, 16(2), 1–8. https://doi.org/10.18637/jss.v016.c02

See also

Examples

prior <- create_t_prior(3, df = c(1,2,5), mu = c(0,1,2), sigma = c(1,2,3))
prior$fn(c(0.25, 0.5, 0.75))
#> [1] -1.000000  1.000000  4.180061