Results from nested sampling runs

The ernest_run object contains the results of a nested sampling run. It has methods for summary, plotting, and for simulating the uncertainty around the run's estimates.

Details

The ernest_run object is returned by running a nested sampling procedure in the ernest package. It can be used for posterior analysis, evidence estimation, and diagnostic plotting.

Slots

n_iter

Total number of iterations performed.

n_points

Number of points drawn into the live set.

n_calls

Number of calls to the log likelihood function.

log_lik

A vector of log likelihood values associated with each point generated during the run.

log_volume

A vector of the estimated prior volumes associated with the removal of each point from the live set.

log_weight

A vector of unnormalized posterior weights for each point.

log_evidence

A vector of log. evidence estimates, generated after the removal of each point.

log_evidence_var

A vector of uncertainty values associated with each entry in log_evidence. 1.

information

A vector of the estimated KL-divergence (or information) between the prior and posterior distributions.

id

The index of each point within the live set.

points

The number of live points associated with each point's removal.

calls

The number of likelihood calls made when generating a replacement live point.

birth

The iteration at which the point was created and added to the live set.

samples

A matrix of the sampled points, expressed in the units of the prior space.

samples_unit

Identical to samples, but expressed in the units of the 0-1 hypercube.

See also