Example Nested Sampling Run with Ernest

This dataset contains the results of a nested sampling run performed on a three-dimensional Gaussian likelihood with independent uniform priors on each parameter.

Usage

data(ernest_run_example)

Format

An ernest_run object containing the results of the nested sampling run. Users can build the run from scratch using vignette("nested-sampling-with-ernest", package = "ernest").

Details

The likelihood used to generate the points is \(MVN(0, \Sigma)\), with each variance in \(\Sigma\) set to 1 and each covariance set to 0.95. The prior for each parameter is uniform on the interval [-10, 10\].

[-10, 10]: R:-10,%2010%5C

See also