Internal: The Ernest Nested Sampler R6 Class

The ernest_sampler is an R6 class that implements the core logic for a nested sampling run.

Details

This class is primarily intended for internal use within the ernest package. Users should interact with nested sampling samplers via the provided S3 generics, rather than calling R6 methods directly.

The R6 interface is documented here for developers and advanced users who want to extend or debug ernest.

Usage

Users are not able to instantiate this class, as it is not exported. Instead, they use the following functions:

To create a new sampler, use nested_sampling().

To compile live points, use compile().

To run nested sampling, use generate().

To access the latest results, use the $run active binding.

Active bindings

niterations

The total number of sampling iterations.

ncalls

The total calls made to the likelihood function.

live_points

A list, containing the matrix of live points currently in the sampler in unit-cube units, and a vector of their associated log likelihood values.

run

The ernest_run object binding the results of the previous sampling runs. Returns NULL if no runs have been performed.

Methods

Public methods

ernest_sampler$new()

ernest_sampler$clear()

ernest_sampler$compile()

ernest_sampler$generate()

ernest_sampler$print()

ernest_sampler$clone()

Method new()

(Internal) Creates a new ernest_sampler.

Usage

ernest_sampler$new(
  log_lik_fn,
  prior,
  sampling,
  n_points,
  first_update,
  update_interval
)

Arguments

log_lik_fn

An ernest_likelihood object.

prior

An ernest_prior object.

sampling

An ernest_sampling object.

n_points

The number of live points.

first_update

The first update interval.

update_interval

The subsequent update interval.

Returns

Itself, invisibly.

Method clear()

(Internal) Clears the previous runs from the sampler, including the sampler's live points.

Usage

ernest_sampler$clear()

Returns

Itself, invisibly.

Method compile()

(Internal) Generates a sample of live points from the prior and validates them.

Usage

ernest_sampler$compile(clear = FALSE)

Arguments

clear

A logical value indicating whether to clear existing points.

Returns

Itself, invisibly.

Method generate()

(Internal) Performs nested sampling until a stopping criterion is met.

Usage

ernest_sampler$generate(
  max_iterations = Inf,
  max_calls = Inf,
  min_logz = 0.05,
  verbose = FALSE
)

Arguments

max_iterations

The maximum number of iterations to perform.

max_calls

The maximum number of calls to the likelihood function.

min_logz

The minimum log-evidence value to achieve.

verbose

Whether to print updates on the sampler's progress.

Returns

Itself, invisibly.

Method print()

(Internal) Prints a brief summary of the sampler.

Usage

ernest_sampler$print(...)

Arguments

...

Ignored.

Returns

Itself, invisibly.

Method clone()

The objects of this class are cloneable with this method.

Usage

ernest_sampler$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.