Specify a likelihood-restricted prior sampler (LRPS)

Choose a method for replacing the worst point in a live set of possible parameter values.

unif_cube() is a region-based sampler, where points are generated by sampling uniformly within the unit hypercube, returning a point when a sampled point exceeds from the prior under a likelihood constraint is to sample randomly from the prior and reject the point if the likelihood constraint is not fulfilled. This method is horribly inefficient in even moderately-large dimensions, but is useful for testing and debugging.

rwmh_cube is a MCMC-based sampler, where points are generated by taking random steps within the unit hypercube.

Usage

unif_cube()

rwmh_cube(steps = 25L, target_acceptance = 0.5)

Arguments

steps

Number of steps to take when generating a proposal point.

target_acceptance

The target acceptance rate for the sampler.

Value

An ernest_sampling object, used by ernest_sampler during its initialization to create an R environment for its LRPS.

Details

Nested sampling first generates a set of independently distributed live points within a prior space, then iteratively improves this set through replacing the point with the worst likelihood with a new, improved point. In ernest, the strategies for evolving these points are collectively called likelihood-restricted prior samplers. At present, ernest offers two LRPS: A naive region-based sampler, which randomly samples from the unit cube, and a more sophisticated random walk Metropolis-Hastings sampler, which takes a random point from the live set and walks it through a number of steps.

When updated during a run (see nested_sampling()), the sampler will adjust its step-size to target the specified acceptance rate.

References

For more on different LRPS, Buchner, J. (2023). Nested Sampling Methods. Statistics Surveys, 17(none), 169–215. https://doi.org/10.1214/23-SS144