Estimate Evidence using a Nested Sampling Run

Computes evidence and related quantities from a nested sampling run, optionally by simulating the volumes of each nested likelihood shell.

Usage

# S3 method for class 'ernest_run'
calculate(x, ..., ndraws = NULL)

Arguments

x

An ernest_run object.

...

These dots are for future extensions and must be empty.

ndraws

An optional positive integer. The number of log-volume sequences to simulate. If equal to zero, no simulations will be made, and a one draw vector of log volumes are produced from the estimates contained in x. If NULL, getOption("posterior.rvar_ndraws") is used (default 4000).

Value

A tibble::tibble(), containing n_iter + n_points rows and several columns:

log_lik: The log-likelihood of the model.

log_volume: The log volume of the prior space.

log_weight: The log weights of the live points.

log_evidence: The log evidence of the model.

log_evidence_err: The standard error of the log evidence (only available when ndraws = 0).

The tibble has the additional class ernest_estimate, which has its own plot method.

Each column is returned as an posterior::rvar() vector.

References

Skilling, J. (2006). Nested Sampling for General Bayesian Computation. Bayesian Analysis, 1(4), 833–859. https://doi.org/10.1214/06-BA127

Examples

# Load an example run
data(example_run)

# View results as a tibble with `ndraws = FALSE` (the default).
calculate(example_run)
#> Nested sampling estimates <ernest_estimate>
#> No. of Simulated Draws: 4000
#> Log. Volume: -16 ± 1.3
#> Log. Evidence: -9 ± 0.05
#> # A tibble: 5,158 × 4
#>       log_lik       log_volume  log_weight log_evidence
#>    <rvar[1d]>       <rvar[1d]>  <rvar[1d]>   <rvar[1d]>
#>  1  -146 ± NA  -0.002 ± 0.0019  -153 ± 1.3  -153 ± 1.26
#>  2  -144 ± NA  -0.004 ± 0.0027  -151 ± 1.3  -151 ± 0.95
#>  3  -139 ± NA  -0.006 ± 0.0033  -146 ± 1.3  -146 ± 1.21
#>  4  -136 ± NA  -0.008 ± 0.0039  -143 ± 1.3  -143 ± 1.09
#>  5  -129 ± NA  -0.010 ± 0.0043  -136 ± 1.3  -136 ± 1.25
#>  6  -126 ± NA  -0.012 ± 0.0048  -134 ± 1.3  -134 ± 1.00
#>  7  -125 ± NA  -0.014 ± 0.0051  -132 ± 1.3  -132 ± 0.87
#>  8  -125 ± NA  -0.016 ± 0.0055  -132 ± 1.3  -131 ± 0.71
#>  9  -118 ± NA  -0.018 ± 0.0058  -126 ± 1.3  -126 ± 1.21
#> 10  -118 ± NA  -0.020 ± 0.0060  -125 ± 1.3  -124 ± 0.83
#> # ℹ 5,148 more rows

# Generate 100 simulated log volume values for each iteration.
calculate(example_run, ndraws = 100)
#> Nested sampling estimates <ernest_estimate>
#> No. of Simulated Draws: 100
#> Log. Volume: -16 ± 1.2
#> Log. Evidence: -8.8 ± 0.12
#> # A tibble: 5,158 × 4
#>       log_lik        log_volume  log_weight log_evidence
#>    <rvar[1d]>        <rvar[1d]>  <rvar[1d]>   <rvar[1d]>
#>  1  -146 ± NA  -0.0021 ± 0.0021  -153 ± 1.2  -153 ± 1.22
#>  2  -144 ± NA  -0.0043 ± 0.0036  -151 ± 1.2  -151 ± 1.06
#>  3  -139 ± NA  -0.0062 ± 0.0041  -146 ± 1.1  -146 ± 1.11
#>  4  -136 ± NA  -0.0084 ± 0.0050  -143 ± 1.2  -143 ± 1.04
#>  5  -129 ± NA  -0.0101 ± 0.0051  -136 ± 1.0  -136 ± 1.01
#>  6  -126 ± NA  -0.0121 ± 0.0058  -134 ± 1.2  -134 ± 0.99
#>  7  -125 ± NA  -0.0140 ± 0.0060  -132 ± 1.3  -132 ± 0.88
#>  8  -125 ± NA  -0.0160 ± 0.0061  -132 ± 1.2  -131 ± 0.70
#>  9  -118 ± NA  -0.0181 ± 0.0063  -126 ± 1.1  -126 ± 1.08
#> 10  -118 ± NA  -0.0200 ± 0.0064  -125 ± 1.0  -124 ± 0.68
#> # ℹ 5,148 more rows