Estimate Evidence using a Nested Sampling Run

Estimate Evidence using a Nested Sampling Run

Usage

# S3 method for class 'ernest_run'
calculate(x, ..., ndraws = FALSE)

Arguments

x

An ernest_run object.

...

These dots are for future extensions and must be empty.

ndraws

A positive integer or a boolean. If FALSE, the log volumes will be derived from their expected values. If not FALSE, the log volumes will be simulated using ndraws samples from each volume's joint distribution.

Value

A tibble, containing run$n_iter + run$n_points rows and the following columns:

log_lik: The log-likelihood of the model.

log_volume: The log volume of the prior space.

log_weight: The log weights of the live points.

log_evidence: The log evidence of the model.

log_evidence.err: The standard error of the log evidence (only available when ndraws = 0).

The tibble has the additional class ernest_estimates, which has its own plot method.

Each column is returned as an posterior::rvar() vector.

Examples

# Load an example run
data(ernest_run_example)

# View results as a tibble with `ndraws = FALSE` (the default).
calculate(ernest_run_example)
#> # A tibble: 10,315 × 5
#>       log_lik   log_volume log_weight log_evidence log_evidence.err
#>    <rvar[1d]>   <rvar[1d]> <rvar[1d]>   <rvar[1d]>       <rvar[1d]>
#>  1  -149 ± NA  -0.001 ± NA  -157 ± NA    -157 ± NA     3.2e-33 ± NA
#>  2  -146 ± NA  -0.002 ± NA  -154 ± NA    -154 ± NA     1.4e-32 ± NA
#>  3  -141 ± NA  -0.003 ± NA  -149 ± NA    -149 ± NA     1.4e-31 ± NA
#>  4  -140 ± NA  -0.004 ± NA  -148 ± NA    -147 ± NA     3.2e-31 ± NA
#>  5  -138 ± NA  -0.005 ± NA  -145 ± NA    -145 ± NA     1.0e-30 ± NA
#>  6  -133 ± NA  -0.006 ± NA  -140 ± NA    -140 ± NA     1.2e-29 ± NA
#>  7  -130 ± NA  -0.007 ± NA  -137 ± NA    -137 ± NA     5.1e-29 ± NA
#>  8  -129 ± NA  -0.008 ± NA  -136 ± NA    -136 ± NA     8.7e-29 ± NA
#>  9  -127 ± NA  -0.009 ± NA  -134 ± NA    -134 ± NA     2.1e-28 ± NA
#> 10  -127 ± NA  -0.010 ± NA  -134 ± NA    -133 ± NA     3.7e-28 ± NA
#> # ℹ 10,305 more rows

# Generate 100 simulated log volume values for each iteration.
calculate(ernest_run_example, ndraws = 100)
#> # A tibble: 10,315 × 4
#>       log_lik          log_volume  log_weight log_evidence
#>    <rvar[1d]>          <rvar[1d]>  <rvar[1d]>   <rvar[1d]>
#>  1  -149 ± NA  -0.00078 ± 0.00073  -157 ± 1.1  -157 ± 1.09
#>  2  -146 ± NA  -0.00190 ± 0.00114  -154 ± 1.1  -154 ± 0.95
#>  3  -141 ± NA  -0.00283 ± 0.00138  -150 ± 1.1  -150 ± 1.04
#>  4  -140 ± NA  -0.00377 ± 0.00160  -148 ± 1.1  -148 ± 0.82
#>  5  -138 ± NA  -0.00479 ± 0.00188  -146 ± 1.4  -145 ± 0.97
#>  6  -133 ± NA  -0.00590 ± 0.00227  -141 ± 1.2  -141 ± 1.20
#>  7  -130 ± NA  -0.00684 ± 0.00241  -138 ± 1.3  -138 ± 1.15
#>  8  -129 ± NA  -0.00770 ± 0.00259  -137 ± 1.4  -137 ± 0.94
#>  9  -127 ± NA  -0.00882 ± 0.00274  -135 ± 1.3  -135 ± 0.90
#> 10  -127 ± NA  -0.00983 ± 0.00288  -134 ± 1.2  -133 ± 0.69
#> # ℹ 10,305 more rows