Estimate Evidence using a Nested Sampling Run

Estimate Evidence using a Nested Sampling Run

Usage

# S3 method for class 'ernest_run'
calculate(x, ..., ndraws = NULL)

Arguments

x

(ernest_run) An ernest_run object.

...

These dots are for future extensions and must be empty.

ndraws

(positive integer or zero, optional) The number of log volume sequences to simulate. If equal to zero, no simulations will be made, and a one draw vector of log volumes are produced from the estimates contained in x. If NULL, getOption("posterior.rvar_ndraws") is used (default 4000).

Value

A tibble, containing run$n_iter + run$n_points rows and the following columns:

log_lik: The log-likelihood of the model.

log_volume: The log volume of the prior space.

log_weight: The log weights of the live points.

log_evidence: The log evidence of the model.

log_evidence_err: The standard error of the log evidence (only available when ndraws = 0).

The tibble has the additional class ernest_estimate, which has its own plot method.

Each column is returned as an posterior::rvar() vector.

Details

Use calculate() to simulate the estimation error nested sampling runs caused by approximating the change in log volume between iterations. Given the use of ordered log likelihood restricted prior sampling, these volumes can be modelled as the order statistics of a uniform random variable. Ernest uses the following distributions to perform this simulation:

Constant number of live points: At each iteration, the point with the lowest likelihood is replaced by a new sample constrained to a higher likelihood region. Under this setup, the shrinkage in prior volume at iteration i can be shown to follow the Beta distribution.

Decreasing number of live points: At the end of the sampling run, the remaining live points are added to the dead point set. This introduces a discrete stepwise behavior in volume shrinkage. Ernest uses a transformation of variables drawn from the exponential distribution to simulate the corresponding changes in volume.

References

See Appendix A6 of Speagle, J. (2020).

Examples

# Load an example run
data(ernest_run_example)

# View results as a tibble with `ndraws = FALSE` (the default).
calculate(ernest_run_example)
#> An <ernest_calculate>: 4000 draws x 10315 iterations

# Generate 100 simulated log volume values for each iteration.
calculate(ernest_run_example, ndraws = 100)
#> An <ernest_calculate>: 100 draws x 10315 iterations