Estimate Evidence using a Nested Sampling Run

Computes evidence and related quantities from a nested sampling run, optionally by simulating the volumes of each nested likelihood shell.

Usage

# S3 method for class 'ernest_run'
calculate(x, ..., ndraws = NULL)

Arguments

x

An ernest_run object.

...

These dots are for future extensions and must be empty.

ndraws

An optional positive integer. The number of log-volume sequences to simulate. If equal to zero, no simulations will be made, and a one draw vector of log-volumes are produced from the estimates contained in x. If NULL, getOption("posterior.rvar_ndraws") is used (default 4000).

Value

A tibble::tibble(), containing n_iter + n_points rows and several columns:

log_lik: The log-likelihood of the model.

log_volume: The log-volume of the prior space.

log_weight: The log weights of the live points.

log_evidence: The log-evidence of the model.

log_evidence_err: The standard error of the log-evidence (only available when ndraws = 0).

The tibble has the additional class ernest_estimate, which has its own plot method.

Each column is returned as an posterior::rvar() vector.

References

Higson, E., Handley, W., Hobson, M., & Lasenby, A. (2019). Nestcheck: Diagnostic Tests for Nested Sampling Calculations. Monthly Notices of the Royal Astronomical Society, 483(2), 2044–2056. https://doi.org/10.1093/mnras/sty3090

Examples

# Load an example run
data(example_run)

# View results as a tibble with `ndraws = FALSE` (the default).
calculate(example_run)
#> evidence estimates <ernest_estimate>
#> 
#> No. of Simulated Draws: 4000
#> Log. Volume: -17 ± 1.3
#> Log. Evidence: -9.2 ± 0.069
#> # A tibble: 10,495 × 4
#>       log_lik         log_volume   log_weight log_evidence
#>    <rvar[1d]>         <rvar[1d]>   <rvar[1d]>   <rvar[1d]>
#>  1  -145 ± NA  -0.00099 ± 0.0010  -152 ± 0.81  -152 ± 0.81
#>  2  -144 ± NA  -0.00199 ± 0.0014  -151 ± 0.80  -150 ± 0.66
#>  3  -137 ± NA  -0.00301 ± 0.0017  -144 ± 0.81  -144 ± 0.81
#>  4  -133 ± NA  -0.00402 ± 0.0020  -140 ± 0.82  -140 ± 0.80
#>  5  -133 ± NA  -0.00504 ± 0.0022  -140 ± 0.79  -139 ± 0.64
#>  6  -131 ± NA  -0.00604 ± 0.0025  -138 ± 0.79  -138 ± 0.61
#>  7  -128 ± NA  -0.00702 ± 0.0027  -136 ± 0.81  -135 ± 0.71
#>  8  -128 ± NA  -0.00804 ± 0.0028  -135 ± 0.81  -135 ± 0.62
#>  9  -127 ± NA  -0.00901 ± 0.0030  -134 ± 0.81  -133 ± 0.58
#> 10  -126 ± NA  -0.01000 ± 0.0031  -134 ± 0.80  -133 ± 0.54
#> # ℹ 10,485 more rows

# Generate 100 simulated log-volume values for each iteration.
calculate(example_run, ndraws = 100)
#> evidence estimates <ernest_estimate>
#> 
#> No. of Simulated Draws: 100
#> Log. Volume: -17 ± 1.4
#> Log. Evidence: -9.1 ± 0.068
#> # A tibble: 10,495 × 4
#>       log_lik         log_volume   log_weight log_evidence
#>    <rvar[1d]>         <rvar[1d]>   <rvar[1d]>   <rvar[1d]>
#>  1  -145 ± NA  -0.00095 ± 0.0010  -152 ± 0.83  -152 ± 0.83
#>  2  -144 ± NA  -0.00192 ± 0.0013  -151 ± 0.77  -150 ± 0.65
#>  3  -137 ± NA  -0.00298 ± 0.0017  -144 ± 0.85  -144 ± 0.85
#>  4  -133 ± NA  -0.00397 ± 0.0020  -140 ± 0.71  -140 ± 0.70
#>  5  -133 ± NA  -0.00477 ± 0.0020  -140 ± 0.78  -139 ± 0.60
#>  6  -131 ± NA  -0.00560 ± 0.0022  -138 ± 0.84  -138 ± 0.60
#>  7  -128 ± NA  -0.00669 ± 0.0023  -136 ± 0.80  -135 ± 0.72
#>  8  -128 ± NA  -0.00751 ± 0.0023  -135 ± 0.84  -135 ± 0.61
#>  9  -127 ± NA  -0.00857 ± 0.0026  -134 ± 0.75  -133 ± 0.59
#> 10  -126 ± NA  -0.00963 ± 0.0029  -134 ± 0.79  -133 ± 0.56
#> # ℹ 10,485 more rows