Prepare an ernest_sampler object for nested sampling

Create a new set of live points/particles for a new nested sampling run, or check the current state of the live points before continuing a previous run.

Usage

# S3 method for class 'ernest_sampler'
compile(object, ..., clear = FALSE)

Arguments

object

An object of class ernest_sampler.

...

These dots are for future extensions and must be empty.

clear

A logical value indicating whether to reload object before creating new live points. If TRUE, former results from the sampler are removed from the object. If FALSE, the sampler will not drop prior results, and continue using live points from the last sampling iteration performed.

Value

object, invisibly.

Details

compile checks whether the live points within a nested sampler are of the expected structure (e.g., the matrix contains the proper dimensions, every live point is described with finite values, and likelihood values are either finite or -Inf). If these checks fail, or if clear = TRUE, the sampler is reset and a new set of live points is generated.

Examples

prior <- create_uniform_prior(n_dim = 2, lower = -1, upper = 1)
ll_fn <- function(x) -sum(x^2)
sampler <- nested_sampling(ll_fn, prior, n_point = 100)

# Add live points to the sampler
compile(sampler)
head(sampler$live_points$unit)
#>             [,1]       [,2]
#> [1,] 0.836140221 0.88949489
#> [2,] 0.909568916 0.07974891
#> [3,] 0.005724923 0.25286010
#> [4,] 0.160214206 0.43794499
#> [5,] 0.968936877 0.23176413
#> [6,] 0.984778127 0.31933229

# Check the status of the sampler with `clear = FALSE`
compile(sampler, clear = FALSE)
head(sampler$live_points$unit)
#>             [,1]       [,2]
#> [1,] 0.836140221 0.88949489
#> [2,] 0.909568916 0.07974891
#> [3,] 0.005724923 0.25286010
#> [4,] 0.160214206 0.43794499
#> [5,] 0.968936877 0.23176413
#> [6,] 0.984778127 0.31933229

# Reset the sampler with new live points with `clear = TRUE`
compile(sampler, clear = TRUE)
head(sampler$live_points$unit)
#>           [,1]      [,2]
#> [1,] 0.7132885 0.3611381
#> [2,] 0.8761787 0.0115667
#> [3,] 0.7810689 0.1651609
#> [4,] 0.1082668 0.5730944
#> [5,] 0.5233944 0.3103647
#> [6,] 0.9491375 0.4526721