Prepare a likelihood function for nested sampling

Creates a modified version of a log. likelihood function that always returns either a finite value or -Inf for each vector of parameters that is provided.

Usage

create_likelihood(fn, ...)

# S3 method for class 'ernest_likelihood'
create_likelihood(
  fn,
  error_action = c("abort", "warn"),
  nonfinite_action = c("warn", "pass", "abort"),
  auto_batch = TRUE,
  ...
)

# S3 method for class '`function`'
create_likelihood(
  fn,
  error_action = c("abort", "warn"),
  nonfinite_action = c("warn", "pass", "abort"),
  auto_batch = TRUE,
  ...
)

Arguments

fn

(uni-variate function) The log-likelihood function (see Details).

...

These dots are for future extensions and must be empty.

error_action

(case-sensitive string) Action to perform once fn throws an error.

"abort": Stop execution and signal an error.

"warn": Issue a warning and replace output with -Inf.

nonfinite_action

(case-sensitive string) Action to perform when fn passes a value that is not a finite double nor -Inf.

"warn": Issue a warning and replace values with -Inf.

"pass": Silently replace values with -Inf.

"abort": Stop execution and signal an error.

auto_batch

(logical) Whether to prepare fn so that it may be called with a matrix of parameter values. If FALSE, its assumed that fn can already produce a vector of likelihood values for a matrix with rows of parameter vectors.

Value

A function with additional class ernest_likelihood. This function will accept vectors of parameters, and matrices of parameters, where each row represents a single parameter vector. ernest_likelihood will always produce either scalar or vector doubles containing finite values or -Inf; values outside this range are either replaced by -Inf or cause errors.

Details

Model likelihoods should be provided as a log density function. It is expected that fn should take in exactly one argument; likelihood functions that take in multiple non-default arguments should be entered as anonymous functions (see rlang::as_function()). See vignettes for an example of entering data through an anonymous function.

It is expected that fn returns a scalar finite values or -Inf for each parameter vector. Use error_action and nonfinite_action to decide on how ernest_likelihood handles errors, non-finite, and non-numeric return values from fn.

As default, auto_batch expects that fn is incapable of handling matrices of parameter values. It resolves this by wrapping fn in a call to base::apply(). Should you have a more efficient implementation of your likelihood function, then consider setting auto_batch to FALSE.

Examples

# A 3D Gaussian likelihood function
n_dim <- 3
sigma <- diag(0.95, nrow = 3)
det_sigma <- determinant(sigma, logarithm = TRUE)$modulus
attributes(det_sigma) <- NULL
prec <- solve(sigma)
log_norm <- -0.5 * (log(2 * pi) * n_dim + det_sigma)

fn <- function(theta) {
  drop(-0.5 * crossprod(theta, crossprod(prec, theta)) + log_norm)
}
log_lik <- create_likelihood(fn)
log_lik(c(0, 0, 0))
#> [1] -2.679876

# As default, `log_lik` will loudly replace non-finite, non-`-Inf` values.
try(log_lik(c(Inf, 0, 0)))
#> Warning: `fn` must return finite numeric values or `-Inf`.
#> ! Replacing `NaN` with `-Inf`.
#> [1] -Inf

# Silence warnings with `nonfinite_action = "pass"`
quiet_lik <- create_likelihood(log_lik, nonfinite_action = "pass")
quiet_lik(c(Inf, 0, 0))
#> [1] -Inf