Prepare a likelihood function for nested sampling

Creates a modified version of a log. likelihood function that always returns either a finite value or -Inf for each vector of parameters that is provided.

Usage

create_likelihood(fn, ...)

Arguments

fn

A function or an object of class ernest_likelihood. If a function, it should take a single numeric vector of parameters and return either a finite log likelihood value or -Inf.

...

These dots are for future extensions and must be empty.

Value

A function with additional class ernest_likelihood. On a success, this function is guaranteed to return a finite value or -Inf. Non-finite, non--Inf values are coerced to -Inf with a warning. Non-numeric returns are reported as errors.

Examples

# A 3D Gaussian likelihood function
n_dim <- 3
sigma <- diag(0.95, nrow = 3)
det_sigma <- determinant(sigma, logarithm = TRUE)$modulus
attributes(det_sigma) <- NULL
prec <- solve(sigma)
log_norm <- -0.5 * (log(2 * pi) * n_dim + det_sigma)

log_lik <- function(theta) {
  drop(-0.5 * crossprod(theta, crossprod(prec, theta)) + log_norm)
}
log_lik(c(0, 0, 0))
#> [1] -2.679876

# ernest catches and warns the user about non-finite, non-`-Inf` values
try(log_lik(c(Inf, 0, 0)))
#> [1] NaN

# ernest will quit when receiving unexpected values from the likelihood function
log_lik_2 <- function(theta) {
 if (theta[1] < 1) "Odd value." else log_lik(theta)
}
log_lik_2(c(1, 0, 0))
#> [1] -3.206191
try(log_lik_2(c(0, 0, 0)))
#> [1] "Odd value."