Perform nested sampling

Execute the nested sampling loop. Iteratively replaces the worst live point in a set with a point drawn from a likelihood-restricted prior sampler, until a provided stopping criterion is met.

Usage

# S3 method for class 'ernest_sampler'
generate(
  x,
  ...,
  max_iterations = NULL,
  max_calls = NULL,
  min_logz = 0.05,
  show_progress = FALSE
)

Arguments

x

(ernest_sampler) An object of class ernest_sampler.

...

Arguments passed on to compile.ernest_run

seed

(integer or NA, optional) Specification for the random number generator.

integer and NULL: Passed to set.seed(). If NULL, this reinitializes the generator as if no seed has yet been set.

NA: Make no changes to the current seed. If compile has been called on object before, then NA will ensure that the seed remain identical between runs.

clear

(boolean) Whether to clear results from previous runs before compiling.

TRUE: Previous results stored in object are removed, and live points are generated and validated.

FALSE: Previous results stored in object are retained, and live points are validated.

max_iterations

(positive integer) The maximum number of iterations to perform. If set to Inf, this stopping criterion is ignored.

max_calls

(positive integer) The maximum number of calls to the likelihood function. If set to Inf, this stopping criterion is ignored.

min_logz

(positive double or zero) The minimum ratio between the current log evidence and remaining log evidence (see Details). If set to zero, this stopping criterion is ignored.

show_progress

Whether to show a simple progress spinner and iteration counter during the run.

Value

An object of class ernest_run, containing the results of the nested sampling run.

Details

At least one of max_iterations, max_calls, or min_logz must represent a non-ignored stopping criterion.

At the iteration \(i\), the remaining log evidence within the prior space bound by a minimum likelihood criterion can be estimated as $$Z^*_i \approx L^{(max)}_i V_i$$ where \(L^{(max)}_i\) is the maximum likelihood of the current live points and \(V_i\) is the estimated remaining volume.

This estimate can be used to define a relative stopping criterion based on the log-ratio of the current estimated evidence \(\hat{Z_i}\) to the remaining evidence, such that $$\delta \log(Z) = \log(\hat{Z_i} - Z^*_i) - \log(\hat{Z_i})$$ By setting min_logz, you can vary the minimum log-ratio at which sampling stops. Once \(\delta \log(Z)\) falls below this value, you can assume that only a negligible fraction of the evidence remains unaccounted for in the final evidence estimates.

If x already contains results from previous calls to generate(), then generate() will ensure that your stopping criterion have not already been surpassed during previous runs.

Control the verbosity of other messages in ernest as it runs with the rlib_message_verbosity enviroment option.

Examples

prior <- create_uniform_prior(n_dim = 2, lower = -1, upper = 1)
ll_fn <- function(x) -sum(x^2)
sampler <- nested_sampling(ll_fn, prior, n_point = 100)
sampler
#> An <ernest_sampler>: 100 points, 2 variables. 

# Stop sampling after a set number of iterations or calls to the lik. func.
generate(sampler, max_iterations = 100)
#> Creating new live points.
#> An <ernest_run>: 100 points, 2 variables
#> • 100 iterations, 168 likelihood calls
#> • Log. Evidence ≈ -0.612 (± 0.124)

# The final number of calls may be larger than `max_calls`, as `generate`
# only checks the number of calls when removing a live point.
generate(sampler, max_calls = 2600)
#> Creating new live points.
#> Error in "lapply(text, glue_cmd, .envir = .envir)": ! Could not evaluate cli `{}` expression: `max_c`.
#> Caused by error in `eval(expr, envir = envir)`:
#> ! object 'max_c' not found

# Use the default stopping criteria
if (FALSE)  generate(sampler)  # \dontrun{}