Prepare a nested sampler to evaluate Bayesian evidence

Initialize a ernest_sampler object to perform nested sampling with a given log-likelihood function, prior distribution, and likelihood-restricted prior specification.

Usage

nested_sampling(
  log_lik,
  prior,
  sampler = rwmh_cube(),
  n_points = 500,
  first_update = NULL,
  update_interval = NULL,
  on_warning = c("abort", "warn"),
  ...
)

Arguments

log_lik

(function) A function that takes a vector of parameters and returns the log-likelihood of the parameters. The function should return a finite value or -Inf for invalid parameters (see create_likelihood()).

prior

(ernest_prior) An object with class ernest_prior, created by create_prior() or its specializations.

sampler

(ernest_sampling) An ernest_sampling object, declaring which likelihood-restricted prior sampler (LRPS) to use.

n_points

(positive integer) The number of live points to use in the nested sampling run.

first_update

(optional positive integer) The number of likelihood calls before adopting LRPS behaviour. If NULL, this is set to n_points * 2.5.

update_interval

(optional positive integer) The number of likelihood calls between updates to the LRPS behaviour. If NULL, this is set to n_points * 1.5.

on_warning

(case-sensitive string) Action to perform when the sanity test throws a warning message (see Details).

"abort": Throw an error.

"warn": Issue a warning and initialize the sampler.

...

Arguments passed on to create_likelihood.function

error_action

(case-sensitive string) Action to perform once fn throws an error.

"abort": Stop execution and signal an error.

"warn": Issue a warning and replace output with -Inf.

nonfinite_action

(case-sensitive string) Action to perform when fn passes a value that is not a finite double nor -Inf.

"warn": Issue a warning and replace values with -Inf.

"pass": Silently replace values with -Inf.

"abort": Stop execution and signal an error.

auto_batch

(logical) Whether to prepare fn so that it may be called with a matrix of parameter values. If FALSE, its assumed that fn can already produce a vector of likelihood values for a matrix with rows of parameter vectors.

Value

An ernest_sampler object, prepared for nested sampling.

Details

This function prepares an ernest_sampler object for performing nested sampling, given a log-likelihood function, prior, and sampler specification. It checks that the provided log-likelihood and prior are compatible by performing a basic sanity check by evaluating the log-likelihood at the center of the prior space. This check ensures that the log-likelihood returns a finite value or -Inf for typical parameter values.

The sanity check may fail even if log_lik is well-specified, for example if the prior's support does not include the test point (the center of the prior hypercube), or if the log-likelihood is only defined for a subset of the parameter space. In such cases, silencing the warning from nested_sampling`` may require you to adjust the prior or adjust the on_warning` argument.

Examples

prior <- create_uniform_prior(n_dim = 2, lower = -1, upper = 1)
ll_fn <- function(x) -sum(x^2)
sampler <- nested_sampling(ll_fn, prior, n_points = 100)
sampler
#> An <ernest_sampler>: 100 points x 0 iter. x 0 lik. calls