Summarise a nested sampling run

Provides a summary of an ernest_run object.

Usage

# S3 method for class 'ernest_run'
summary(object, ...)

Arguments

object

(ernest_run) An object of class ernest_run.

...

These dots are for future extensions and must be empty.

Value

An object of class summary.ernest_run, a list with:

n_iter: Number of iterations (number of dead points).

n_points: Number of live points at the end of the run.

log_volume, log_evidence, log_evidence_err: The final estimates of the quantities performed by the run that generated object.

run, A tibble with n_iter + n_points rows, containing the vectors call, log_lik, log_volume, log_weight, log_evidence, log_evidence_err, and information.

See also

Examples

# Load an example run
data(ernest_run_example)

# Summarise the run and view a tibble of its results.
run_sm <- summary(ernest_run_example)
run_sm
#> 
#> ── Nested Sampling Results from <ernest_run> ───────────────────────────────────
#> No. Points: 1000
#> No. Iterations: 9315
#> No. Lik. Calls: 232875
#> Log. Evidence: -8.978 (± 0.08203)
#> 
#> ── Weighted Posterior Distribution 
#> # A tibble: 3 × 7
#>   variable               mean  median    sd   mad    q5   q95
#>   <chr>                 <dbl>   <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 Uniform(-10, 10)    0.0119   0.0222 0.969 0.984 -1.55  1.61
#> 2 Uniform(-10, 10).1 -0.0209  -0.0149 0.972 0.965 -1.64  1.55
#> 3 Uniform(-10, 10).2  0.00132  0.0104 0.986 1.00  -1.61  1.59
run_sm$run
#> # A tibble: 10,315 × 7
#>     call log_lik log_volume log_weight log_evidence log_evidence_err information
#>    <int>   <dbl>      <dbl>      <dbl>        <dbl>            <dbl>       <dbl>
#>  1    25   -149.     -0.001      -148.        -157.         3.17e-33   -1.01e-62
#>  2    50   -146.     -0.002      -145.        -154.         1.43e-32   -2.04e-61
#>  3    75   -141.     -0.003      -140.        -149.         1.41e-31   -1.99e-59
#>  4   100   -140.     -0.004      -139.        -147.         3.25e-31   -1.05e-58
#>  5   125   -138.     -0.005      -136.        -145.         1.01e-30   -1.03e-57
#>  6   150   -133.     -0.006      -131.        -140.         1.17e-29   -1.36e-55
#>  7   175   -130.     -0.007      -128.        -137.         5.06e-29   -2.56e-54
#>  8   200   -129.     -0.008      -128.        -136.         8.74e-29   -7.65e-54
#>  9   225   -127.     -0.009      -125.        -134.         2.14e-28   -4.60e-53
#> 10   250   -127.     -0.01       -125.        -133.         3.66e-28   -1.34e-52
#> # ℹ 10,305 more rows